requirements.txt
numpy >= 1.11.2
pandas >= 0.19.1
requests >= 2.11.1
#scrapy >= 1.3.2
websocket-client == 0.40.0